Separated classes used by tests
